Move Engine to its own file.
It makes more sense for us to not pile too much functionality into the
mod.rs file.
